Refactored pages to containers
