+++ src/containers/index.js
@@ -1,2 +1,6 @@
 export { default as QuestionPageContainer } from './QuestionPageContainer.js';
 export { default as QuestionBasePageContainer } from './QuestionBasePageContainer.js';
+export { default as RedefinePasswordPageContainer } from './RedefinePasswordPageContainer.js';
+export { default as UserPasswordProfileContainer } from './UserPasswordProfileContainer.js';
+export { default as UserProfilePageContainer } from './UserProfilePageContainer.js';
+export { default as ForgotPasswordPageContainer } from './ForgotPasswordPageContainer.js';
